requirement.txt
amqp==5.1.1
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==15.1.0
cors==1.0.1
cryptography==41.0.4
daphne==4.0.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.2
django-cors-headers==4.2.0
django-crontab==0.7.1
django-import-export==3.3.1
django-tinymce==3.6.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.4
drf-spectacular-sidecar==2023.10.1
et-xmlfile==1.1.0
filelock==3.12.4
future==0.18.3
gevent==23.9.1
greenlet==3.0.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kombu==5.3.2
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
prompt-toolkit==3.0.39
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.2.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
requests-file==1.5.1
rest-framework-simplejwt==0.0.2
rpds-py==0.10.3
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
tldextract==3.6.0
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.6
uWSGI==2.0.23
vine==5.0.0
wcwidth==0.2.8
xlrd==2.0.1
xlwt==1.3.0
zope.event==5.0
zope.interface==6.0
